harmonisation des Controller et des Resources
